Wrote investments by redux
